tools/airtest-assets/requirements.txt
numpy<2.0
opencv-python-headless>=4.8,<5.0
Pillow>=10.0
six>=1.16,<2.0
